working with preorder different

+++ include/VertexEdge.h
@@ -26,12 +26,12 @@ private:
      * @brief If vertex was visited
      */
     bool _visited = false;
-    
+
     /**
      * @brief If vertex is processing (used for DAGs)
      */
     bool _processing = false;
-    
+
     /**
      * @brief Vertex indegree (used for topsort)
      */
@@ -57,21 +57,21 @@ public:
 
     /**
      * @brief Get the vertex id
-     * 
+     *
      * @return int id
      */
     int getId() const;
 
     /**
      * @brief Get the adjacency list of edges
-     * 
+     *
      * @return std::vector<Edge*> adjacencyList
      */
     std::vector<Edge *> getAdj() const;
 
     /**
      * @brief If the vertex was visited
-     * 
+     *
      * @return true Vertex was visited
      * @return false Vertex was not visited
      */
@@ -79,7 +79,7 @@ public:
 
     /**
      * @brief If the vertex is processing
-     * 
+     *
      * @return true Vertex is processing
      * @return false Vertex is not processing
      */
@@ -87,7 +87,7 @@ public:
 
     /**
      * @brief Get vertex indegree
-     * 
+     *
      * @return unsigned int indegree
      */
     unsigned int getIndegree() const;
@@ -101,43 +101,43 @@ public:
 
     /**
      * @brief Get vertex path
-     * 
+     *
      * @return Edge* path
      */
     Edge* getPath() const;
 
     /**
      * @brief Get incomming edges to the vertex
-     * 
+     *
      * @return std::vector<Edge*> incommingEdges
      */
     std::vector<Edge *> getIncomming() const;
 
     /**
      * @brief Set vertex id
-     * 
+     *
      * @param int id
      */
     void setId(int id);
 
     /**
      * @brief Set vertex to visited/unvisited
-     * 
-     * @param visited 
+     *
+     * @param visited
      */
     void setVisited(bool visited);
 
     /**
      * @brief Set vertex to (not) processing
-     * 
-     * @param processing 
+     *
+     * @param processing
      */
     void setProcessing(bool processing);
 
     /**
      * @brief Set vertex indegree
-     * 
-     * @param indegree 
+     *
+     * @param indegree
      */
     void setIndegree(unsigned int indegree);
 
@@ -150,14 +150,14 @@ public:
 
     /**
      * @brief Set path to vertex
-     * 
-     * @param path 
+     *
+     * @param path
      */
     void setPath(Edge* path);
 
     /**
      * @brief Add an edge with vertex as origin
-     * 
+     *
      * @param dest Destination Vertex
      * @param weight Edge weight
      * @return Edge* New edge
@@ -166,7 +166,7 @@ public:
 
     /**
      * @brief Remove an edge with the vertex as origin
-     * 
+     *
      * @param destId Id of the destination vertex
      * @return true Vertex was removed
      * @return false Vertex does not exist
@@ -175,7 +175,7 @@ public:
 };
 
 /**
- * @brief Represents the path between vertexes 
+ * @brief Represents the path between vertexes
  */
 class Edge {
 private:
@@ -183,12 +183,12 @@ private:
      * @brief Destination vertex
      */
     Vertex* _dest;
-    
+
     /**
      * @brief Distance between vertexes
      */
     double _weight;
-    
+
     /**
      * @brief Origin vertex
      */
@@ -205,53 +205,53 @@ private:
     double _flow;
 
 public:
-    Edge(Vertex* origin, Vertex* dest, int weight);
+    Edge(Vertex* origin, Vertex* dest, double weight);
 
     /**
      * @brief Get the destination vertex
-     * 
+     *
      * @return Vertex* destination
      */
     Vertex* getDest() const;
 
     /**
      * @brief Get the edge's weight
-     * 
+     *
      * @return int weight
      */
     double getWeight() const;
 
     /**
      * @brief Get the origin vertex
-     * 
+     *
      * @return Vertex* origin
      */
     Vertex* getOrigin() const;
 
     /**
      * @brief Get reverse edge
-     * 
+     *
      * @return Edge* reverse
      */
     Edge* getReverse() const;
 
     /**
      * @brief Get the edge's flow
-     * 
+     *
      * @return int Edge flow
      */
     double getFlow() const;
 
     /**
      * @brief Set reverse edge
-     * 
+     *
      * @param reverse reverse
      */
     void setReverse(Edge* reverse);
 
     /**
      * @brief Set the flow
-     * 
+     *
      * @param flow Edge flow
      */
     void setFlow(double flow);
